Remove Typesense stubs and flatten search abstraction
The original PR shipped a SearchProvider interface plus a Typesense stub
in src/frontend/src/scripts/search/ so the Pagefind -> Typesense swap
would be a one-line change. We're no longer doing that migration, so
the abstraction layer has no remaining justification.

* Delete src/frontend/src/scripts/search/typesense-provider.ts
* Delete src/frontend/src/scripts/search/SearchProvider.ts
* Delete src/frontend/src/scripts/search/index.ts
* Delete src/frontend/src/scripts/search/pagefind-provider.ts
* Inline the Pagefind logic and types into a single
  src/frontend/src/scripts/search.ts that exports
  searchAspireDocs(query, limit).
* webmcp.ts now imports searchAspireDocs directly; drop the stale
  "Pagefind today, Typesense later" comment and the provider-selector
  indirection.

No behavior change for the WebMCP search-aspire-docs tool. Pagefind is
the only backend now, and the unavailable-fallback contract is unchanged.
